# docs/openapi.json
{
  "info": {
    "description": "REST API extension for Burp Suite Professional that enables programmatic access to proxy traffic, scanner functionality, scope management, and collaborative testing workflows. Features advanced response analysis, real-time proxy interception, and enhanced security testing capabilities using Montoya API 2025.8+.",
    "title": "burp2api - Burp Suite REST API Extension",
    "version": "1.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/examples": {
      "get": {
        "description": "Interactive request examples for the documented endpoints",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Interactive request examples for the documented endpoints",
        "tags": [
          "Documentation"
        ]
      }
    },
...
    "/session/current": {
      "get": {
        "description": "Get current session tag with statistics and usage information",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get current session tag with statistics and usage information",
        "tags": [
          "Session"
        ]
      }
    },
    "/session/help": {
      "get": {
        "description": "Complete documentation for session management API",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Complete documentation for session management API",
        "tags": [
          "Session"
        ]
      }
    },
    "/session/history": {
      "get": {
        "description": "Get complete session history and statistics",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get complete session history and statistics",
        "tags": [
          "Session"
        ]
      }
    },
    "/session/new": {
      "post": {
        "description": "Generate a new session tag and apply it (optional: prefix parameter)",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Generate a new session tag and apply it (optional: prefix parameter)",
        "tags": [
          "Session"
        ]
      }
    },
    "/session/tag": {
      "post": {
        "description": "Update session tag with auto-generation support (auto-generates if empty)",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Update session tag with auto-generation support (auto-generates if empty)",
        "tags": [
          "Session"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local Burp API Server",
      "url": "http://localhost:7850"
    }
  ],
...
}